Add LC_VERSION_MASK facility to querylocale(3).
Allow user programs to ask for a version string for the
components of a locale_t.  The structure of the version
string is undefined, but can be used to detect changes
in the locale's definition.

The intended use-case is databases that use libc collations
to implement indexes.  When collation definitions change,
database indexes are frequently corrupted.  By exposing
the CLDR version string, databases can detect the change
and issue an error or force an index rebuild.

For now only LC_COLLATE components have a way to report
their version.  The LC_COLLATE file format is backwards
and forwards compatible.  Where previously 24 bytes held
"BSD 1.0\n" followed by NUL characters, there are now 12
bytes for the file format identifier and then 12 bytes for
the data version string.