Allow LIO_READV and LIO_WRITEV in lio_listio().
Allow multiple vector IOs to be started with one system call.  The
aio_readv() and aio_writev() functions already used much of the same
code as lio_listio() under the covers.  This commit exposes the formerly
internal-only LIO_READV and LIO_WRITEV opcodes to user space.

The new opcode are only visible to user space if __BSD_VISIBLE is
defined, being extensions to POSIX, just like the aio_readv() and
aio_writev() functions.

The other obvious candidate is LIO_SYNC, but that's left for later.