kqueue: Add umtx (futex) support.
Allow user space mutex wakeups to be reported as kevents, so
that they can be multiplexed with other kinds of kevent.

The new filter EVFILT_UMTX fires immediately if the memory
pointed to by "ident" does not contain the value in "data".
Otherwise, it fires on the next qualifying call to
_umtx_op(UMTX_OP_WAKE).

For examples of possible uses, this provides Capsicum-friendly
alternative to signals, or as a building block for IPC that is more
efficient than pipes.

XXX WIP