kqueue: Add EVFILT_USERMEM.
Provide an umtx facility (AKA futexes) via kqueue.  This allows
processes to wake each other up using only a piece of shared memory, and
to multiplex with other kinds of kevents.

The new filter EVFILT_USERMEM fires immediately if the memory pointed to
by "ident" does not contain the value in "data".  Otherwise, it fires on
the next NOTE_TRIGGER or _umtx_op(UMTX_OP_WAKE) for the same address.

XXX WIP