nfsclient: Fix readdir() cache concurrency.
A loop that reads a directory while someone else changes the directory
could previously see incoherent results, for example missing entries
that are in fact still there.  That happened when synchronization was
lost between the position in the directory stream (represented by a
cookie issued by the server) and the position in the local buffer cache.

This doesn't happen on OpenBSD's implementation, because every
getdirentries() call makes an RPC call to the server, using a cookie
that should usually still be accepted on typical implementations.

This commit keeps FreeBSD's much faster cache for unchanging
directories, but falls back to OpenBSD's approach for already-started
readdir() loops when the directory contents changes.  New readdir()
loops are able to use the cache.

To achieve this, we now expose the server's cookies directly to user
space, like OpenBSD does, so that successive getdirentries() calls
move through the directory stream from cookie to cookie, instead of
using linear offsets into the cache.  This is important, because nothing
else in the universe remembers where a given caller's readdir() loop was
up to.  Not the kernel, not the NFS server.  With that change made, now
we just need to be careful to invalidate everything reliably when a
change is detected, and provide a fallback back when cookies are not
recognized.